Added opportunities saved layout
