+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Layout/ignitus-OpportunitiesSaved/constants.ts
@@ -1,0 +1,8 @@
+export const opportunity = [
+{ 'title':'PhD Project', 'description':'Church and civil courts in late medieval and early modern Scotland', 'subtitle':'By The University of Melbourne', 'data':'../../../ignitus-Assets/ignitus-Images/img-Png/universityOfMelbourne.png' },
+{ 'title':'Internship', 'description':'Church and civil courts in late medieval and early modern Scotland', 'subtitle':'By The University of Melbourne', 'data':'../../../ignitus-Assets/ignitus-Images/img-Png/universityOfPortsmouth.png' },
+{ 'title':'Funded Research', 'description':'Church and civil courts in late medieval and early modern Scotland', 'subtitle':'By The University of Melbourne', 'data':'../../../ignitus-Assets/ignitus-Images/img-Png/tohokuUniversity.png' },
+{ 'title':'Internship', 'description':'Church and civil courts in late medieval and early modern Scotland', 'subtitle':'By The University of Melbourne', 'data':'../../../ignitus-Assets/ignitus-Images/img-Png/murdochUniversity.png' },
+{ 'title':'Internship', 'description':'Church and civil courts in late medieval and early modern Scotland', 'subtitle':'By The University of Melbourne', 'data':'../../../ignitus-Assets/ignitus-Images/img-Png/microsoft.png' },
+{ 'title':'Professor\'s Assistant', 'description':'Church and civil courts in late medieval and early modern Scotland', 'subtitle':'By The University of Melbourne', 'data':'../../../ignitus-Assets/ignitus-Images/img-Png/universityOfNewcastle.png' },
+];
