+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Organisms/ignitus-OpportunitySaved/Components/OpportunitySaved.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
 import * as S from '../styles';
-import universityOfMelbourne from '../../../ignitus-Assets/ignitus-Images/img-Png/universityOfMelbourne.png'
 import { Description } from '../types';
-import { opportunity } from '../constants';
 import {withErrorBoundary} from '../../../../ignitus-ErrorHandlingComponents/errorBoundary';
 
 export const OpportunitySavedCard = ({title, description, subtitle, data}: Description) => (
   <S.OpportunityContainer>
     <S.ImageDescriptionContainer>
       <S.ImageContainer>
-        <img src={universityOfMelbourne} alt={title} height="120" />
+        <img src={data} alt={title} height="120" />
       </S.ImageContainer>
       <S.DescriptionContainer>
         <S.OpportunityTitle>{title}</S.OpportunityTitle>
@@ -22,14 +20,4 @@ export const OpportunitySavedCard = ({title, description, subtitle, data}: Descr
   </S.OpportunityContainer>
 );
 
-const OpportunitySavedLayout = () => (
-  <S.OpportunityList>
-    <S.OpportunitiesSavedGrid>
-    {opportunity.map(({title, description, subtitle, data}: Description) => (
-        <OpportunitySavedCard title={title} description={description} subtitle={subtitle} data={data} />
-    ))}
-    </S.OpportunitiesSavedGrid>
-  </S.OpportunityList>
-);
-
-export default withErrorBoundary(OpportunitySavedLayout);
+export default withErrorBoundary(OpportunitySavedCard);
