+++ src/ignitus-UserInterfaceKit/Components/Organisms/interfaceCardLayout/style.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { flexibleRowDiv, flexibleColDiv } from '../../../../ignitus-Shared/ignitus-DesignSystem/shared';
+import { flexibleColDiv } from '../../../../ignitus-Shared/ignitus-DesignSystem/shared';
 
 export const CardLayoutContainer = styled.nav`
   padding: 2rem;
